Move theme check in LIA, and show splash while its evaluating
Also do not let login screen start before this process is done
